pyproject.toml
[project]
name = "fastapi_tortoise_mysql"
version = "0.0.1"
description = "fastapi tortoise-orm with mysql"
authors = [
    { name = "Wu Clan", email = "jianhengwu0407@gmail.com" },
]
dependencies = [
    "aerich==0.7.2",
    "asgiref==3.8.1",
    "bcrypt==4.1.3",
    "cryptography==42.0.7",
    "email_validator==2.1.1",
    "fast-captcha==0.2.1",
    "fastapi[all]==0.111.0",
    "fastapi-limiter==0.1.6",
    "fastapi-pagination==0.12.24",
    "loguru==0.7.2",
    "pwdlib==0.2.1",
    "path==16.14.0",
    "pre-commit==3.7.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "redis[hiredis]==5.0.4",
    "tortoise-orm[asyncmy]==0.24.0",
    "msgspec>=0.18.6",
    "phonenumbers>=8.13.51",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm.dev-dependencies]
lint = [
    "ruff>=0.9.2",
]
deploy = [
    "gunicorn==22.0.0",
    "wait-for-it>=2.2.2",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
lint = "pre-commit run --all-files"
